Reintroduce Einschreibung.Intervall but do not allow to enroll students in short blocks.
This comes as of the newest requirements. Students should not be able to enroll in any subblocks. They may however speak with supervision to change their enrollment mid-block. To make this trackable, i've reintroduced the field.

To prevent misuse of this to circumvent the required-categories rule by showing up somewhere only to ask for a change 5 min later but still fulfilling the category, a category only counts as fulfilled after being there for at least a full block.